OneDrive/Desktop/HACK-A-THON 1.0/requirements.txt
Flask==2.3.3
pymongo==4.6.0
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2
gspread==6.1.4
oauth2client==4.1.3
google-api-python-client==2.160.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
PyDrive2==1.21.3
ollama==0.6.0
PyPDF2==3.0.1
